Move logger to pkg/logger
